Refactor: Move SetViperStructDefaults to utils
